requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
pydantic>=1.9.0
requests>=2.28.0
openai>=1.0.0
python-dotenv>=0.21.0
typing-extensions>=4.0.0
sqlalchemy>=2.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0